fix(US Kia): Correct temperature setting
In the US Kia vehicle and UVO app, the next step below 62 is "LOW" and the next step above 82 is "HIGH". So if you try to set the climate to an actual number lower than 62 or higher than 82, it causes an error that leads to the engine and the rear defroster coming on, but the AC does not. I don't know of an easy way to change the service call itself to accept for numbers but also the words "LOW" and "HIGH", so adding the logic check to the API corrects for that.