BREAKING CHANGE: reduce number of exports
Confirmed that kia_uvo only needs the remaining imports. Kept
WindowRequestOptions even if it is not used yet.

Remove the API implementation from being exported since the user should
use VehicleManager().

The goal is to make internal refactoring safer by reducing the risk for
future breaking changes.

This make it possible to reduce code duplication in the future.

Add virtualenv items to .gitignore.