Rename *APIUSA to *ApiUSA
The fact that the USA one had "API" ALL CAPS but all others had "Api"
really triggered me. This should not be a breaking change because
clients use VehiculeManager but there's still a risk.

exceptions.py also use API all caps but that would break clients so this
is not changed.