fix(CA): fixed the cloudflare 403 issue
Updated the login API to v2 and changed the 'From' header from SPA to CWP based on captured traffic between the browser and the mybluelink.ca site. Dropped legacy cipher support.

It seems to be working on my HAOS but might need more testing. It appears that Cloudflare believes part of the request contains a string that triggered its security solution.