fix(CA): fixed the cloudflare 403 issue (#724)
Updated the login API to v2 and changed the 'From' header from SPA to
CWP based on captured traffic between the browser and the mybluelink.ca
site. Dropped legacy cipher support. V2 login has a different structure
for accesstoken and refresh token, therefore, updated the way to retrive
the tokens.

It seems to be working on my HAOS but might need more testing. It
appears that Cloudflare believes part of the request contains a string
that triggered its security solution.

Below SS were from the api calling mybluelink.ca

---


![image](https://github.com/user-attachments/assets/6520b5fd-f958-4130-8ba1-189a05dd3f9f)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>