Fix Canada/Newfoundland time zone support
Canadian users can configure a timezone in their car to one of the
Canadian timezones, it does not have to support arbitary others. It also
turns out that Newfoundland is an odd timezone with half an hour diff.

Refactor the code to check for one of the known timezones instead of
inventing a dummy "VEHICLETIME" one. This allows dropping a dependency
on tzoffset from the external dateutil library.

Drop unused `self.vehicle_timezone`, removed since commit c885a81f2ae6.